[script] split t1-helper to two main binary
