+++ script/default.nix
@@ -9,63 +9,70 @@
 }:
 
 let
-  self = stdenv.mkDerivation rec {
-    name = "t1-helper";
+  mkHelper = { moduleName, scriptSrc, outName }:
+    let
+      self = stdenv.mkDerivation rec {
+        name = "t1-${moduleName}-script";
 
-    src = with lib.fileset; toSource {
-      root = ./.;
-      fileset = unions [
-        ./src
-        ./build.sc
-      ];
-    };
+        src = with lib.fileset; toSource {
+          root = ./.;
+          fileset = unions [
+            scriptSrc
+            ./build.sc
+          ];
+        };
 
-    passthru.millDeps = fetchMillDeps {
-      inherit name;
-      src = with lib.fileset; toSource {
-        root = ./.;
-        fileset = unions [
-          ./build.sc
+        passthru.millDeps = fetchMillDeps {
+          inherit name;
+          src = with lib.fileset; toSource {
+            root = ./.;
+            fileset = unions [
+              ./build.sc
+            ];
+          };
+          millDepsHash = "sha256-J8bBgM/F+8x8EQ1DR6Va/ZY2hnsjkkzk4a+ctDMKK3k=";
+        };
+
+        passthru.withLsp = self.overrideAttrs (old: {
+          nativeBuildInputs = old.nativeBuildInputs ++ [
+            metals
+            # Metals require java to work correctly
+            mill.passthru.jre
+          ];
+
+          shellHook = ''
+            grep -q 'hardfloat' build.sc \
+              && echo "Please run nix develop in ./script directory instead of project root" \
+              && exit 1
+
+            mill -i mill.bsp.BSP/install 0
+          '';
+        });
+
+        nativeBuildInputs = [
+          mill
+          graalvm-ce
+
+          makeWrapper
+          passthru.millDeps.setupHook
         ];
+
+        buildPhase = ''
+          echo "Building JAR"
+          mill -i ${moduleName}.assembly
+          echo "Running native-image"
+          native-image --no-fallback -jar out/${moduleName}/assembly.dest/out.jar "$name.elf"
+        '';
+
+        installPhase = ''
+          mkdir -p "$out"/bin
+          cp "$name.elf" "$out"/bin/"${outName}"
+        '';
       };
-      millDepsHash = "sha256-J8bBgM/F+8x8EQ1DR6Va/ZY2hnsjkkzk4a+ctDMKK3k=";
-    };
-
-    passthru.withLsp = self.overrideAttrs (old: {
-      nativeBuildInputs = old.nativeBuildInputs ++ [
-        metals
-        # Metals require java to work correctly
-        mill.passthru.jre
-      ];
-
-      shellHook = ''
-        grep -q 'hardfloat' build.sc \
-          && echo "Please run nix develop in ./script directory instead of project root" \
-          && exit 1
-
-        mill -i mill.bsp.BSP/install 0
-      '';
-    });
-
-    nativeBuildInputs = [
-      mill
-      graalvm-ce
-
-      makeWrapper
-      passthru.millDeps.setupHook
-    ];
-
-    buildPhase = ''
-      echo "Building JAR"
-      mill -i assembly
-      echo "Running native-image"
-      native-image --no-fallback -jar out/assembly.dest/out.jar "$name.elf"
-    '';
-
-    installPhase = ''
-      mkdir -p "$out"/bin
-      cp "$name.elf" "$out"/bin/"$name"
-    '';
-  };
+    in
+    self;
 in
-self
+{
+  t1-helper = mkHelper { moduleName = "emu"; scriptSrc = ./emu; outName = "t1-helper"; };
+  ci-helper = mkHelper { moduleName = "ci"; scriptSrc = ./ci; outName = "ci-helper"; };
+}
